Clarify cooker reset wording new start-rd-on-exit method for Configurator

+++ retrodeck.sh
@@ -47,9 +47,12 @@ https://retrodeck.net
       cli_compress_all_games "$2"
       ;;
     --configurator*)
-      launched_from_cli="true"
       sh /app/tools/configurator.sh
-      exit
+      if [[ $(configurator_generic_question_dialog "RetroDECK Configurator" "Would you like to launch RetroDECK after closing the Configurator?") == "false" ]]; then
+        exit
+      else
+        shift
+      fi
       ;;
     --reset-emulator*)
       echo "You are about to reset one or more RetroDECK emulators."
@@ -104,16 +107,16 @@ if [ -f "$lockfile" ]; then
     echo "Config file's version is $version but the actual version is $hard_version"
     if grep -qF "cooker" <<< $hard_version; then # If newly-installed version is a "cooker" build
       cooker_base_version=$(echo $hard_version | cut -d'-' -f2)
-      choice=$(zenity --icon-name=net.retrodeck.retrodeck --info --no-wrap --ok-label="Upgrade" --extra-button="Don't Upgrade" --extra-button="Fresh Install" \
+      choice=$(zenity --icon-name=net.retrodeck.retrodeck --info --no-wrap --ok-label="Upgrade" --extra-button="Don't Upgrade" --extra-button="Full Wipe and Fresh Install" \
       --window-icon="/app/share/icons/hicolor/scalable/apps/net.retrodeck.retrodeck.svg" \
       --title "RetroDECK Cooker Upgrade" \
-      --text="You appear to be upgrading to a \"cooker\" build of RetroDECK.\n\nWould you like to perform the standard post-update process, skip the post-update process or remove ALL existing RetroDECK data to start from a fresh install?\n\nPerforming the normal post-update process multiple times may lead to unexpected results.")
+      --text="You appear to be upgrading to a \"cooker\" build of RetroDECK.\n\nWould you like to perform the standard post-update process, skip the post-update process or remove ALL existing RetroDECK folders and data (including ROMs and saves) to start from a fresh install?\n\nPerforming the normal post-update process multiple times may lead to unexpected results.")
       rc=$? # Capture return code, as "Yes" button has no text value
       if [[ $rc == "1" ]]; then # If any button other than "Yes" was clicked
         if [[ $choice == "Don't Upgrade" ]]; then # If user wants to bypass the post_update.sh process this time.
           echo "Skipping upgrade process for cooker build, updating stored version in retrodeck.cfg"
           set_setting_value $rd_conf "version" "$hard_version" retrodeck # Set version of currently running RetroDECK to updated retrodeck.cfg
-        elif [[ $choice == "Fresh Install" ]]; then # Remove all RetroDECK data and start a fresh install
+        elif [[ $choice == "Full Wipe and Fresh Install" ]]; then # Remove all RetroDECK data and start a fresh install
           echo "Removing RetroDECK data and starting fresh"
           rm -rf /var
           rm -rf "$HOME/retrodeck"
